requirements.txt
pyyaml>=6.0
tqdm>=4.64
torch>=2.0
transformers>=4.35
sentence-transformers>=2.2
chromadb>=0.4
bert-score>=0.3
nltk>=3.8
pillow>=9.0
datasets>=2.14
